Refactor kafka package
