+++ src/main/kotlin/no/nav/syfo/App.kt
@@ -18,9 +18,10 @@ import no.nav.syfo.client.pdfgen.PdfGenClient
 import no.nav.syfo.client.veiledertilgang.VeilederTilgangskontrollClient
 import no.nav.syfo.client.wellknown.getWellKnown
 import no.nav.syfo.melding.MeldingService
-import no.nav.syfo.melding.kafka.*
 import no.nav.syfo.melding.kafka.config.KafkaBehandlerDialogmeldingSerializer
+import no.nav.syfo.melding.kafka.dialogmelding.launchKafkaTaskDialogmeldingFraBehandler
 import no.nav.syfo.melding.kafka.legeerklaring.launchKafkaTaskLegeerklaring
+import no.nav.syfo.melding.kafka.producer.DialogmeldingBestillingProducer
 import no.nav.syfo.melding.status.kafka.launchKafkaTaskDialogmeldingStatus
 import org.apache.kafka.clients.producer.KafkaProducer
 import org.slf4j.LoggerFactory
