+++ src/test/kotlin/no/nav/syfo/melding/cronjob/MeldingFraBehandlerCronjobSpek.kt
@@ -8,8 +8,8 @@ import io.mockk.verify
 import kotlinx.coroutines.runBlocking
 import no.nav.syfo.melding.database.getMeldingerForArbeidstaker
 import no.nav.syfo.melding.database.updateInnkommendePublishedAt
-import no.nav.syfo.melding.kafka.KafkaMeldingFraBehandlerProducer
-import no.nav.syfo.melding.kafka.PublishMeldingFraBehandlerService
+import no.nav.syfo.melding.kafka.producer.KafkaMeldingFraBehandlerProducer
+import no.nav.syfo.melding.kafka.producer.PublishMeldingFraBehandlerService
 import no.nav.syfo.testhelper.ExternalMockEnvironment
 import no.nav.syfo.testhelper.UserConstants
 import no.nav.syfo.testhelper.createMeldingerFraBehandler
